# Performance: prevent blowup in normalization