@pombredanne 2022-05-02
I am merging this in https://github.com/bastikr/boolean.py/pull/107 ... so I am closing this here